Split instances
